third_party/blink/web_tests/wpt_internal/route/parsing/serialize-001.html

<!DOCTYPE html>
<title>Serialization of @route</title>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://wicg.github.io/declarative-partial-updates/css-route-matching/">
<style class="test" id="sheet1">
  @route (at: r66) { }
</style>
<style class="test" id="sheet2">
  @route (from: r66) or (to: blah) { }
</style>
<style class="test" id="sheet3">
  @route ((from: r66) or (to: blah)) { }
</style>
<style class="test" id="sheet4">
  @route (Attention! This is General Enclosed speaking) { }
</style>
<style class="test" id="sheet5">
  @route (at: urlpattern("/r66/")) { }
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  let expectations = [ '@route (at: r66)',
                       '@route (from: r66) or (to: blah)',
                       '@route ((from: r66) or (to: blah))',
                       '@route (Attention! This is General Enclosed speaking)',
                       '@route (at: urlpattern("/r66/"))',
                     ];
  document.querySelectorAll(".test").forEach((elm) => {
    test(()=> {
      assert_greater_than(expectations.length, 0);
      let expectation = expectations[0];
      expectations.shift();
      assert_equals(elm.sheet.cssRules.length, 1);
      const everything = elm.sheet.cssRules[0].cssText;
      const index = everything.indexOf("{");
      assert_not_equals(index, -1);
      assert_equals(everything.substring(0, index).trim(), expectation);
    }, "Serialize route in " + elm.id);
  });
</script>
